[db] Move redis publisher to gitpod-db
