+++ components/gitpod-db/package.json
@@ -31,7 +31,9 @@
     "slugify": "^1.6.5",
     "the-big-username-blacklist": "^1.5.2",
     "typeorm": "0.2.38",
-    "ioredis": "^5.3.2"
+    "ioredis": "^5.3.2",
+    "ioredis-mock": "^8.7.0",
+    "prom-client": "^13.2.0"
   },
   "devDependencies": {
     "@testdeck/mocha": "^0.3.3",
