+++ components/gitpod-db/src/index.ts
@@ -33,3 +33,5 @@ export * from "./personal-access-token-db";
 export * from "./typeorm/entity/db-personal-access-token";
 export * from "./linked-in-profile-db";
 export * from "./redis/client";
+export * from "./redis/publisher";
+export * from "./redis/metrics";
