--- components/ws-manager-bridge/src/redis/publisher.ts
@@ -1,69 +1,0 @@
-/**
- * Copyright (c) 2023 Gitpod GmbH. All rights reserved.
- * Licensed under the GNU Affero General Public License (AGPL).
- * See License.AGPL.txt in the project root for license information.
- */
-
-import { inject, injectable } from "inversify";
-import { log } from "@gitpod/gitpod-protocol/lib/util/logging";
-import { Metrics } from "../metrics";
-import {
-    HeadlessUpdatesChannel,
-    PrebuildUpdatesChannel,
-    RedisHeadlessUpdate,
-    RedisPrebuildUpdate,
-    RedisWorkspaceInstanceUpdate,
-    WorkspaceInstanceUpdatesChannel,
-} from "@gitpod/gitpod-protocol";
-import { Redis } from "ioredis";
-
-@injectable()
-export class RedisPublisher {
-    constructor(@inject(Redis) private readonly redis: Redis, @inject(Metrics) private readonly metrics: Metrics) {}
-
-    async publishPrebuildUpdate(update: RedisPrebuildUpdate): Promise<void> {
-        log.debug("[redis] Publish prebuild udpate invoked.");
-
-        let err: Error | undefined;
-        try {
-            const serialized = JSON.stringify(update);
-            await this.redis.publish(PrebuildUpdatesChannel, serialized);
-            log.debug("[redis] Succesfully published prebuild update.", update);
-        } catch (e) {
-            err = e;
-            log.error("[redis] Failed to publish prebuild update.", e, update);
-        } finally {
-            this.metrics.reportUpdatePublished("prebuild", err);
-        }
-    }
-
-    async publishInstanceUpdate(update: RedisWorkspaceInstanceUpdate): Promise<void> {
-        let err: Error | undefined;
-        try {
-            const serialized = JSON.stringify(update);
-            await this.redis.publish(WorkspaceInstanceUpdatesChannel, serialized);
-            log.debug("[redis] Succesfully published instance update.", update);
-        } catch (e) {
-            err = e;
-            log.error("[redis] Failed to publish instance update.", e, update);
-        } finally {
-            this.metrics.reportUpdatePublished("workspace-instance", err);
-        }
-    }
-
-    async publishHeadlessUpdate(update: RedisHeadlessUpdate): Promise<void> {
-        log.debug("[redis] Publish headless udpate invoked.");
-
-        let err: Error | undefined;
-        try {
-            const serialized = JSON.stringify(update);
-            await this.redis.publish(HeadlessUpdatesChannel, serialized);
-            log.debug("[redis] Succesfully published headless update.", update);
-        } catch (e) {
-            err = e;
-            log.error("[redis] Failed to publish headless update.", e, update);
-        } finally {
-            this.metrics.reportUpdatePublished("headless", err);
-        }
-    }
-}
